Accounts (module implementation and side work)

+++ src/account/account.service.ts
@@ -1,7 +1,14 @@
 import { Injectable } from '@nestjs/common'
 
 import { Types } from 'mongoose'
+import {
+  ConflictError,
+  NotFoundError,
+  ValidationError,
+} from '../common/errors/errors'
+import { UserService } from '../user/user.service'
 import { AccountDatabaseService } from './account.database.service'
+import { AccountType } from './account.types'
 import { AccountQueryParamsDto } from './dto/account-query-params.dto'
 import { CreateAccountDto } from './dto/create-account.dto'
 import { UpdateAccountDto } from './dto/update-account.dto'
@@ -11,9 +18,24 @@ import type { Account } from './schemas/account.schema'
 export class AccountService {
   constructor(
     private readonly accountDatabaseService: AccountDatabaseService,
+    private readonly userService: UserService,
   ) {}
 
   async createAccount(createAccountDto: CreateAccountDto): Promise<Account> {
+    const { userId, accountType, isSavings } = createAccountDto
+
+    if (accountType === AccountType.LOAN && isSavings) {
+      throw new ValidationError('Loan account type cannot be marked as savings')
+    }
+
+    const isUserExist = await this.userService.isUserExistByQuery({
+      _id: userId,
+    })
+
+    if (!isUserExist) {
+      throw new NotFoundError(`User with userId ${userId} not found`)
+    }
+
     return await this.accountDatabaseService.createAccount(createAccountDto)
   }
 
@@ -22,6 +44,16 @@ export class AccountService {
   }
 
   async getAccountsByUser(options: AccountQueryParamsDto): Promise<Account[]> {
+    const { userId } = options
+
+    const isUserExist = await this.userService.isUserExistByQuery({
+      _id: userId,
+    })
+
+    if (!isUserExist) {
+      throw new NotFoundError(`User with userId ${userId} not found`)
+    }
+
     return await this.accountDatabaseService.getAccountsByUser(options)
   }
 
@@ -29,10 +61,26 @@ export class AccountService {
     id: Types.ObjectId,
     updateAccountDto: UpdateAccountDto,
   ): Promise<Account> {
+    const { accountType } = await this.getAccountById(id)
+
+    if (accountType === AccountType.LOAN && updateAccountDto.accountType) {
+      throw new ValidationError(
+        'Changing the account type for loan accounts is not possible',
+      )
+    }
+
     return await this.accountDatabaseService.updateAccount(id, updateAccountDto)
   }
 
   async deleteAccount(id: Types.ObjectId): Promise<void> {
+    const { balance } = await this.getAccountById(id)
+
+    if (balance !== 0) {
+      throw new ConflictError(
+        'Account with a non-zero balance cannot be deleted. You must clear the account first by transferring funds from it',
+      )
+    }
+
     return await this.accountDatabaseService.deleteAccount(id)
   }
 }
